Add files via upload
Employees are said to be the most valueable resource in any organisation. An employee leaving the organization can have a very negative impactive impact in the day to day functioning of it. With the use of analytics and machine learning we will try to unliean patterns that cause employees to leave an organisation.

The IBM HR Analytics Attrition Dataset provides comprehensive information about employee demographicsjob roles, and satisfaction levels, as well as indicators of attrition. The dataset comprises information such as employees' age, gender, education level, job role, monthly income, work-life balance, and years of experience with the company, among others.

The dataset's primary objective is to understand the factors contributing to employee attrition and develop predictive models to anticipate and mitigate attrition risks. By analyzing this dataset, HR professionals and business leaders can gain valuable insights into employee retention strategies and improve overall workforce management practices.It comprises 1470 observations and 35 features.